Sort images for an interview

This workhorse function for sort_images() works within an interview folder of the exported images and copies images to their destination directory based on the product-unit found in the file name.

sort_images_obs(dir_in, unit_labels, dir_out)

Arguments

dir_in
Character. Folder for a single interview whose images to sort

unit_labels
Named numeric vector.

dir_out
Character. Root folder of destination folder for sorted images.

Details

First, it determines where files should go by:

Creating a list of files in the folder

Parsing the file name into product name and unit code components

Constructs a file name that includes the file's interview__key

Constructing a destination file path

Then, it moves the file to its destination directory by:

Checking whether the directory exists

If not, issuing a warning

If so, copying the file